+++ src/App.jsx
@@ -17,6 +17,7 @@ import ControlePlacarNormal from './pages/ControlePlacarNormal';
 import Tabela from './pages/Tabela';
 import ControleTabela from './pages/ControleTabela';
 import TabelaCompacta from './pages/TabelaCompacta';
+import ControleTabelaCompacta from './pages/ControleTabelaCompacta';
 import MataMata from './pages/MataMata';
 import ControleMataMata from './pages/ControleMataMata';
 import FasesFinais from './pages/FasesFinais';
@@ -70,6 +71,10 @@ export default function App() {
           <Route path="/tabela" element={<Tabela />} />
           <Route path="/tabela/controle" element={<ControleTabela />} />
           <Route path="/tabela-compacta" element={<TabelaCompacta />} />
+          <Route
+            path="/tabela-compacta/controle"
+            element={<ControleTabelaCompacta />}
+          />
           <Route path="/mata-mata" element={<MataMata />} />
           <Route path="/mata-mata/controle" element={<ControleMataMata />} />
           <Route path="/fases-finais" element={<FasesFinais />} />